Enable linter analysis cache build phase and run the build tagger before linting
- Linting is now done in two phases: an initial analysis cache build phase,
  and then the actual linting phase, to decrease the memory footprint of
  linting.
