+++ .github/workflows/ci.yml
@@ -99,8 +99,9 @@ jobs:
       - name: Lint
         env:
           GOLANGCI_LINT_CACHE: ${{ steps.go_cache.outputs.analysis_cache }}
-          SKIP_LINTER_ANALYSIS: true
-          RUN_BUILDTAGGER: false
+          SKIP_LINTER_ANALYSIS: false
+          RUN_BUILDTAGGER: true
+          GOGC: "50"
         run: make lint
 
   check-diff:
